alloauto-segmentation-training/requirements.txt
torch
numpy
pandas
transformers
scikit-learn
onnxruntime
pytorch-crf
datasets
python-docx
matplotlib
sentencepiece
accelerate
tensorboard
